added comments to Sketch component and outsourced sketch function to utils folder
